requirements.txt
flask==2.0.3
gunicorn==20.1.0
flask-socketio==5.3.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.70.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
python-dotenv==0.21.0
requests==2.28.2
firebase-admin==6.1.0
eventlet==0.24.1 
python-engineio==4.3.4
python-socketio==5.7.2